requirements.txt
backcall==0.2.0
gunicorn==20.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
dataclasses==0.6
decorator==5.1.1
dnspython==2.2.1
Flask==2.0.3
idna==3.3
importlib-metadata==4.8.3
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
lxml==4.9.1
flask_cors
MarkupSafe==2.0.1
mysql-connector==2.2.9
pandas
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.31
Pygments==2.13.0
pymongo==4.1.1
pytz==2022.2.1
requests==2.27.1
selenium==3.141.0
six
soupsieve==2.3.2.post1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.0.3
wincertstore==0.2
zipp==3.6.0
Pillow
pytube
gridfs-fuse